requirements.txt
asgiref==3.11.0
Django==6.0
django-cors-headers==4.2.0
django-filter==25.2
django-jazzmin==3.0.1
djangorestframework==3.15.1
Pillow==11.0.0
PyJWT==2.10.1
python-dotenv==1.0.1
sqlparse==0.5.5
tzdata==2025.3
python-decouple==3.8
drf-nested-routers>=0.95.0